fix group checks on room toolbox items
